Integers as default (#648)
This adds to CN an integer mode that interprets C integers using mathematical integers (with associated range constraints). The mode is not yet dynamically switchable, but the default from now on. Non-linear arithmetic is handled using uninterpreted functions (as previously prototyped by @moratorium08, and as in a previous CN version) And, @rbanerjee20 's integer-mode fixes for Fulminate.

------------------------------------------------

* fix some tests

* fixes for integer-mode and test updates

* add bw_{and/or/xor}_uf

* dune fmt

* typing rules for shifts in integer mode, Hiro's optimisation

* more

* test output fixes

* remove in_z_range Loc logic

* Revert "remove in_z_range Loc logic"

This reverts commit 229a4a98a655117901af5a5dce362315f280a18c.

* fix int/pointer cast rules for integer-mode

* arithmetic fixes and additions

* more

* wib

* run CI tests with `--always-interp` and adjust test outputs
accordingly

* CLZ, CTZ, FFS, FLS

* update tests

* fix test outputs

* more

* dune fmt

* test fixes

* fix, as found by Dhruv

* more, fixes

* fix

* fixes

* fix tests, switch VIP tests to cvc5

* test fixes

* dune fmt

* use the 'fix-integers' CN-tutorial branch while the PR is in-progress

* wib

* add missing terms as unsupported to cn-to-coq

* dune fmt

* disable Rocq workflow

* update cn-seq-test-gen tests

* Fix function that assumed bitvector type where it could be an integer

* Add main function to new div.c test so Fulm can run

* Try to make mask_ptr.c spec integer-friendly -- no longer failing CN typechecking

* Mark tests with assertions over too-large integer literals as buggy

* tentatively use 'fix-integers' CN tutorial branch for Fulminate CI

* Fix syntax error in proof CI file

* Fix div.c for proof mode

* Update mask_ptr.c.verify to reflect in-progress change

* once more

* and again

* update cvc5 download script

* disable, for now, cvc5 tutorial CI (using only z3)

* point fulminate and proof ci at regular cn-tutorial branch again

---------

+++ tests/cn-seq-test-gen/src/runway.pass.c
@@ -1,6 +1,6 @@
-/*@ function (i32) INACTIVE () { 0i32 } @*/
+/*@ function (integer) INACTIVE () { 0 } @*/
 static int c_INACTIVE() { return 0; }
-/*@ function (i32) ACTIVE () { 1i32 } @*/
+/*@ function (integer) ACTIVE () { 1 } @*/
 static int c_ACTIVE() { return 1; }
 
 struct State {
@@ -17,17 +17,17 @@ function (boolean) valid_state (struct State s) {
      (s.ModeD == INACTIVE() || s.ModeD == ACTIVE()) &&
      (s.ModeA == INACTIVE() || s.ModeD == INACTIVE()) &&
 
-     (s.W_A >= 0i32 && s.W_D >= 0i32) &&
-     (0i32 <= s.Runway_Time && s.Runway_Time <= 5i32) &&
-     (0i32 <= s.Plane_Counter && s.Plane_Counter <= 3i32) &&
+     (s.W_A >= 0 && s.W_D >= 0) &&
+     (0 <= s.Runway_Time && s.Runway_Time <= 5) &&
+     (0 <= s.Plane_Counter && s.Plane_Counter <= 3) &&
 
      (s.ModeA == INACTIVE() && s.ModeD == INACTIVE()
-        implies s.Plane_Counter == 0i32) &&
-     (s.Runway_Time > 0i32
+        implies s.Plane_Counter == 0) &&
+     (s.Runway_Time > 0
         implies (s.ModeA == ACTIVE() || s.ModeD == ACTIVE())) &&
 
-     (s.Plane_Counter > 0i32 && s.ModeA == ACTIVE() implies s.W_D > 0i32) &&
-     (s.Plane_Counter > 0i32 && s.ModeD == ACTIVE() implies s.W_A > 0i32)
+     (s.Plane_Counter > 0 && s.ModeA == ACTIVE() implies s.W_D > 0) &&
+     (s.Plane_Counter > 0 && s.ModeD == ACTIVE() implies s.W_A > 0)
 }
 @*/
 struct State init()
@@ -40,13 +40,13 @@ struct State init()
 }
 struct State increment_Plane_Counter(struct State s)
     /*@ requires valid_state(s);
-                 0i32 <= s.Plane_Counter;
-                 s.Plane_Counter <= 2i32;
+                 0 <= s.Plane_Counter;
+                 s.Plane_Counter <= 2;
                  s.ModeA == ACTIVE() || s.ModeD == ACTIVE();
-                 s.ModeA == ACTIVE() implies s.W_D > 0i32;
-                 s.ModeD == ACTIVE() implies s.W_A > 0i32;
+                 s.ModeA == ACTIVE() implies s.W_D > 0;
+                 s.ModeD == ACTIVE() implies s.W_A > 0;
         ensures  valid_state(return);
-                 s.Plane_Counter == return.Plane_Counter - 1i32;
+                 s.Plane_Counter == return.Plane_Counter - 1;
                  s.Runway_Time == return.Runway_Time;
                  s.ModeA == return.ModeA;
                  s.ModeD == return.ModeD;
@@ -61,7 +61,7 @@ struct State increment_Plane_Counter(struct State s)
 struct State reset_Plane_Counter(struct State s)
     /*@ requires valid_state(s);
         ensures  valid_state(return);
-                 return.Plane_Counter == 0i32;
+                 return.Plane_Counter == 0;
                  s.Runway_Time == return.Runway_Time;
                  s.ModeA == return.ModeA;
                  s.ModeD == return.ModeD;
@@ -78,7 +78,7 @@ struct State increment_Runway_Time(struct State s)
     /* --BEGIN-- */
     /*@ requires valid_state(s);
                  0i32 <= s.Runway_Time;
-                 s.Runway_Time <= 4i32;
+                 s.Runway_Time <= 4;
                  s.ModeA == ACTIVE() || s.ModeD == ACTIVE();
         ensures  valid_state(return);
                  s.Plane_Counter == return.Plane_Counter;
@@ -95,7 +95,7 @@ struct State reset_Runway_Time(struct State s)
     /* --BEGIN-- */
     /*@ requires valid_state(s);
         ensures  valid_state(return);
-                 return.Runway_Time == 0i32;
+                 return.Runway_Time == 0;
                  s.ModeA == return.ModeA;
                  s.ModeD == return.ModeD;
                  s.W_A == return.W_A;
@@ -111,17 +111,17 @@ struct State reset_Runway_Time(struct State s)
 struct State arrive(struct State s)
     /* --BEGIN-- */
     /*@ requires valid_state(s);
-                 s.ModeA == ACTIVE() && s.W_A >= 1i32;
-                 s.Plane_Counter <= 2i32;
+                 s.ModeA == ACTIVE() && s.W_A >= 1;
+                 s.Plane_Counter <= 2;
         ensures  valid_state(return);
                  s.Runway_Time == return.Runway_Time;
                  s.ModeA == return.ModeA;
                  s.ModeD == return.ModeD;
                  s.W_D == return.W_D;
-                 s.W_D == 0i32
+                 s.W_D == 0
                    implies s.Plane_Counter == return.Plane_Counter;
-                 s.W_D > 0i32
-                   implies s.Plane_Counter == return.Plane_Counter - 1i32;
+                 s.W_D > 0
+                   implies s.Plane_Counter == return.Plane_Counter - 1;
     @*/
     /* --END-- */
 {
@@ -135,8 +135,8 @@ struct State arrive(struct State s)
 struct State depart(struct State s)
     /* --BEGIN-- */
     /*@ requires valid_state(s);
-                 s.ModeD == ACTIVE() && s.W_D >=1i32;
-                 s.Plane_Counter <= 2i32;
+                 s.ModeD == ACTIVE() && s.W_D >=1;
+                 s.Plane_Counter <= 2;
         ensures  valid_state(return);
                   s.Runway_Time == return.Runway_Time;
                   s.ModeA == return.ModeA;
@@ -156,7 +156,7 @@ struct State switch_modes(struct State s)
     /* --BEGIN-- */
     /*@ requires valid_state(s);
                  s.ModeA == ACTIVE() || s.ModeD == ACTIVE();
-                 s.Plane_Counter == 0i32;
+                 s.Plane_Counter == 0;
         ensures  valid_state(return);
                  return.ModeA == ACTIVE() || return.ModeD == ACTIVE();
                  return.ModeA == s.ModeD;
@@ -192,13 +192,13 @@ struct State switch_modes(struct State s)
 struct State tick(struct State s)
     /* --BEGIN-- */
     /*@ requires valid_state(s);
-                 (i64) s.Plane_Counter < 2147483647i64;
-                 (i64) s.W_A < 2147483647i64;
-                 (i64) s.W_D < 2147483647i64;
+                 s.Plane_Counter < 2147483647;
+                 s.W_A < 2147483647;
+                 s.W_D < 2147483647;
         ensures valid_state(return);
-                (s.W_A > 0i32 && s.W_D == 0i32 && s.Runway_Time == 0i32
+                (s.W_A > 0 && s.W_D == 0 && s.Runway_Time == 0
                    implies return.ModeA == ACTIVE());
-                (s.W_D > 0i32 && s.W_A == 0i32 && s.Runway_Time == 0i32
+                (s.W_D > 0 && s.W_A == 0 && s.Runway_Time == 0
                    implies return.ModeD == ACTIVE());
     @*/
     /* --END-- */
