+++ tests/cn/array_shift_void.error.c
@@ -5,9 +5,9 @@ void *_malloc(size_t size)
 /*@ trusted;
 ensures
   take A = Alloc(return);
-  A.base == (u64) return;
+  A.base == (integer) return;
   A.size == size;
-  take B = each(u64 i; i < size) {W<unsigned char>(array_shift(return, i))};
+  take B = each(integer i; i < size) {W<unsigned char>(array_shift(return, i))};
 @*/
 {
     return malloc(size);
