+++ tests/cn/int_to_ptr.c
@@ -1,7 +1,7 @@
 void* cast(unsigned long long addr)
 /*@
 ensures
-    addr == 0u64 && is_null(return) || addr != 0u64 && has_alloc_id(return) && (u64) return == addr;
+    addr == 0 && is_null(return) || addr != 0 && has_alloc_id(return) && (integer) return == addr;
 @*/
 {
     return (void*)addr;
@@ -11,5 +11,5 @@ int main()
 {
     int x = 0;
     void* p = cast((unsigned long long)&x);
-    /*@ assert ((u64) &x == 0u64 || has_alloc_id(p) && (u64) p == (u64) &x); @*/
+    /*@ assert ((integer) &x == 0 || has_alloc_id(p) && (integer) p == (integer) &x); @*/
 }
